Remove MTK includes and flags with no effect.
The MTK variables added to LINUXINCLUDE and KBUILD_* are usually not
initialized when they are assigned, so the assignment is useless. Only
when the Makefile is included from other Makefiles after the variables
were set the assignments do something, and only the one for
LINUXINCLUDE, as the variables assigned to KBUILD_* are not set
anywhere. Even when LINUXINCLUDE is set the assigned value is not
needed, so the assignments can be safely removed.

Also, there are no "*_mtk_cust.mak", so the include is also removed.