Remove superfluous include and forward declaration from "mach/env.h".
The "partition_define.h" header is only needed for "dum-char" and
"mmc-host" drivers, so it is no longer included in the general
"mediatek/platform/mt6589/kernel/core/include/mach/env.h" header.

In the same way, the "eMMC_rw_x" function is only needed for "dum-char"
driver, so it was also removed from the general "mach/env.h" header.