Use partition definition files from their original directory.
The "partition_define.h" and "pmt.h" files are in the
"mediatek/custom/ahong89_wet_jb2/common/" directory, but during the
build they were expected to be in the
"mediatek/custom/out/ahong89_wet_jb2/common/". Unlike other custom
files, they were not automatically copied to the destination directory
by the build system, which caused the build to fail (unless they were
manually copied before the build, or unless ptgen.pl was run before the
build, as that tool also performs the copy). Instead of automatically
copying them, the build system was fixed to use the files from their
original directory.

The "dum-char" and "mmc-host" drivers are the only elements that need
"partition_define.h" and "pmt.h", so the original directory is now
included through "ccflags-y" in their Makefiles. This follows the
approach used in the official Android kernel for MediaTek devices, in
which MediaTek related headers are included through "ccflags-y" in
"drivers/misc/mediatek/Makefile.custom" which, in turn, is included by
other Makefiles (although, in that case, all the headers are included,
instead of just those needed for each element).

The original directory depends on the MediaTek project being built, so
the configuration option "ARCH_MTK_PROJECT" was added to Kconfig to be
able to set the proper directory in the Makefiles (again, following the
official Android kernel for MediaTek devices).

Finally, as the partition definition files are now explicitly included
from their original directory in the Makefiles of those elements that
need them, there is no need to add "$(MTK_ROOT_CUSTOM_OUT)/common" nor
"$(MTK_ROOT_CUSTOM)/$(MTK_PROJECT)/common" to LINUXINCLUDE.