Add proper build system for "mediatek/custom/".
Instead of through standard Makefiles, the "mediatek/custom/" directory
was built by copying some of the files of that directory to
"mediatek/custom/out/" and then building those copied files using make
rules generated dynamically on each make run based on the contents of
"mediatek/custom/out/".

Now, that custom build system was replaced with standard Makefiles,
although only for those elements needed to build a Fairphone 1 kernel;
the rest of elements, like drivers for hardware not present in the
Fairphone 1, were not taken into account in the added Makefiles (as they
are useless for the purpose of this kernel tree).

The elements that are taken into account are either built no matter
what, or they are built or not depending on the chosen Kconfig options.
In order to handle all the needed elements some options were added to
Kconfig, based on the options from ProjectConfig.mk used by the custom
build system to select the files to copy, and also on the Kconfig
options used in the official Android kernel for MediaTek devices.

How header files from "mediatek/custom/" were found during the build was
also based on the official Android kernel for MediaTek devices; instead
of adding their directories to LINUXINCLUDE (like it was done through
the generated rules in the custom build system) they are added to
ccflags-y in the Makefile for the elements that need them as
"-I<directory>". This differs slightly from the official Android kernel
for MediaTek devices, though, as in that case all the directories that
contain custom headers are added to ccflags-y (through the inclusion of
"drivers/misc/mediatek/Makefile.custom"), while here only those
directories specifically needed for each element are explicitly added.

Finally, besides copying to "mediatek/custom/out/" the
"mediatek/custom/" files really used to build the kernel, the build
system also copied the "mediatek/custom/" files related to Android HAL,
which are not needed to build the kernel. Therefore, the copy of Android
HAL files when building the kernel was removed.