Generate MediaTek configuration files in output directory.
The mediatek-configs target generates a kernel configuration file by
merging the kernel configuration files disseminated through
"mediatek/config/*/autoconfig/kconfig" directories. That file was
generated under the source tree, but now it is generated under the
kernel output directory (like the ".config" file that it will be copied
to).

The output directory for all the configuration and custom files
generated by the MediaTek build system when the kernel was
built was set too to the kernel output directory. After adding a proper
build system for "mediatek/custom/" it was not really needed, as the
merged configuration file was the only one being created, but it was
modified anyway just for consistency.