Generate MediaTek make rules only for kconfig.
From all the make rules generated dynamically by
"mtk.config.generate-rules", only those used to, in turn, generate
"$(MTK_ROOT_GEN_CONFIG)/kconfig" are needed to build the kernel.
Therefore, the call to "mtk.config.generate-rules" was replaced with a
more specific one that only generates the kconfig rules.

Besides, a rule to create the "$(MTK_ROOT_GEN_CONFIG)" directory had to
be added too, as it is not automatically created by the kconfig rules.
That directory is listed as a dependency of the generated target for the
kconfig file, and it has no relevance for the "mediatekconfig" target,
so it was removed from its dependencies.