Fix cleaning of MediaTek kernel files.
The custom "clean" target for MediaTek kernel files is not needed, as
after the changes to the MediaTek build system the "clean" target from
the standard kernel build system can remove almost all the generated
files. The only one that needs to be explicitly removed is
"$(MTK_ROOT_GEN_CONFIG)/kconfig", but it can be accomplished just by
adding it to the MRPROPER_FILES variable (or in CLEAN_FILES, but being a
configuration file, even if a temporal one not really used in the build
itself, it seemed better to be removed by "mrproper").