Fix building on arbitrary output directory.
The relative path using ancestors in an include of "xlog_logger.c"
imposed a specific location for the output directory of the build. To
make possible to use arbitrary output directories, the path of the
include was replaced with the same path used in the xlog driver from the
official Android kernel for MediaTek devices, which does not contain
ancestors.