Add defconfig for Fairphone 1 recovery.
The standard fp1_defconfig generates a kernel that, for recoveries like
ClockworkMod, is too large to fit along the recovery ramdisk in the
recovery partition (as it only has 6MB).

A specific configuration based on the standard fp1_defconfig but that
builds a smaller kernel targeted at recovery images is introduced.

This configuration enables XZ compression of the kernel image instead of
GZIP (CONFIG_KERNEL_XZ) and uses the "-Os" flag instead of "-O2" when
building the kernel (CONFIG_CC_OPTIMIZE_FOR_SIZE). The downside of those
changes is slightly slower startup and slightly slower general
performance, although both are acceptable for a recovery kernel.

The rest of the configuration is the same found in fp1_defconfig.
Further changes not suitable for a standard kernel but acceptable in a
recovery kernel could be added in the future if the kernel size needs to
be reduced even more.