Set partition layout for original Fairphone 1.
The partition definitions were set based on the values returned by
"/proc/dumchar_info" in a running official kernel (which are not exactly
the same that should appear in the partition definitions; the addresses
are slightly changed when output, as can be seen in "dumchar.c") and
ptgen.pl source code. The previous partition definitions were probably
intended for a Fairphone 1 with unified partitions, although I do not
have one to check it.

The size for the FAT partition was not really needed to be set, as
dum-char module ignores it and always assigns the remaining space from
the start of the partition. Anyway, it was just set for consistency.

On the other hand, the proper value for MTK_SHARED_SDCARD must be set in
ProjectConfig.mk, as whether it is defined or not affects how dum-char
and the USB gadget are built.

Only those files used by the kernel were updated; other files in the
MediaTek tree that store partition layout information but are not used
by the kernel were just ignored.