Fix building with GCC 4.7.
In GCC 4.7, an error is thrown if the body of functions marked as
"always_inline" is not available. The GCC optimizations defined in
"kernel/include/linux/compiler-gcc.h" force the "always_inline" property
for every "inline" function in ARM (as the CONFIG_* variables checked
are not available in ARM), so the building fails due to some inlined
functions not having their body available.

Based on the code for the affected files in the official Android kernel
for MediaTek devices, the problem was fixed just removing the offending
"inline" declarations.