Prevent Android Build System from reading Android.mk files.
The Android.mk files in the kernel tree are not needed to build the
kernel, but some of them clash with other Android.mk used elsewhere
in the Android tree. Therefore, an empty Android.mk file was added to
prevent the Android Build System from finding other deeper Android.mk
files.