Honor PVR build variant defined in kernel configuration.
The PVR driver provides a configuration choice to build it in release or
debug mode. However, in practice the options set by that choice were
ignored; the driver was built in release or debug mode based on the
value of the TARGET_BUILD_VARIANT environment variable, which is set by
the general Android build system.

Although it could be useful to build the driver in debug mode when
Android is built in engineer mode, overriding the kernel configuration,
the problem is that the user libraries built in release mode can not be
used with a kernel driver built in debug mode (among other reasons,
because the ioctl request codes used in the PVR driver change depending
on the build mode). The user libraries are proprietary and only those
built in release mode are provided with the device, so even if Android
and the kernel itself are built in engineer mode the PVR driver must be
built in release mode, as it would be useless otherwise.

Note that the official repository for the Fairphone kernel seems to
contain PVR user libraries built in debug mode. However, given the
directory they are in, the fact that they are proprietary and the lack
of clear licensing regarding reuse or redistribution it is safer to
assume that they can not be used freely.

Therefore, now the TARGET_BUILD_VARIANT is no longer taken into account
regarding the selection of the build mode for the PVR driver and the
CONFIG_PVR_BUILD_RELEASE_MT6589 and CONFIG_PVR_BUILD_DEBUG_MT6589 kernel
configuration options are used instead.