Add command line parameters expected to be set by the bootloader.
When Android boots up, the "init" executable processes the kernel
command line and sets "ro.boot.XXX" Android properties based on the
"androidboot.XXX" parameters of the command line. Some of those
properties are used, in turn, to set other read-only properties, like
"ro.bootmode" or "ro.serialno". However, if the kernel command line does
not contain the associated parameters, the properties are set to their
default value. Being read-only, they can not be modified once set, thus
keeping that default value until the next boot.

The bootloader of the Fairphone 1 does not set any "androidboot.XXX"
parameter. In some cases it does not really matter if the associated
property is set to its default value. But, in other cases, the
properties have an important role in the boot process (like
"ro.bootmode") or even in the system itself (like "ro.serialno").
Therefore, the kernel was modified, when possible, to provide the needed
"androidboot.XXX" parameters that should have come from the bootloader.

Unfortunately, only "androidboot.mode" can be set in this way; when the
kernel command line is tuned "serial_number" was not yet set, and its
initialization depends on functions that can not be called yet either,
so "androidboot.serialno" can not be set to a unique identifier.

Instead of tuning the kernel command line, another approach, that also
seems the only one valid in order to set the serial number of the
device, would be to modify the code of the "init" executable itself (the
boot mode can be read from "/sys/class/BOOT/BOOT/boot/boot_mode", and
the serial number from "/sys/sys_info/serial_number"). However, even if
it is not a complete solution, the kernel itself seems like a more
appropriate place to fix the command line parameters provided by the
bootloader.

Of course, this change to add the command line parameters is needed only
for the standard "init" executable; the MediaTek version of the "init"
executable is able to set the properties to the appropriate values even
if there are no "androidboot.XXX" parameters in the command line.
Anyway, as this change simply adds one of those parameters it should
still be compatible with the MediaTek version of the "init" executable.