Fix reading and writing beyond the device size in dumchar driver.
The dumchar driver offers access to several areas of the main disk as
character devices. To do this, in some cases its operations delegate on
the operations of the main disk. As EMMC device data is read from or
written to the main disk, the number of bytes to be read or written must
be limited based on the size of the EMMC device; otherwise, if the
number of bytes to be read or written exceeded the remaining size of the
EMMC device from the current position, part of the data would be read or
written outside the area of the desired EMMC device and from/to the area
in the main disk of the next EMMC device.

This change implicitly fixes the use of the "read" and "write" functions
from user space with devices managed by the dumchar driver, as instead
of stopping and returning EOF when the end of the device file was
reached they just kept reading and writing more data.