Fix uevents sent when physically disconnecting the USB cable.
When the Android Composite USB Driver is disabled/disconnected, the
upper Android layers (like UsbDeviceManager) expect a
"USB_STATE=DISCONNECTED" uevent to be sent. However, when the USB cable
is physically disconnected, the MediaTek modified driver sends a
MediaTek specific "USB_STATE=HWDISCONNECTED" uevent instead. The driver
was modified to also send the default expected uevent before the
MediaTek specific uevent.

Although the Android Composite USB Driver found in the official Android
kernel for MediaTek devices does not include any MediaTek specific code
(it is the same as the driver in the official Android kernel), for the
time being all the MediaTek specific code was not removed to keep this
kernel compatible with the official Fairphone ROM. Therefore, the fix is
just an addition to the MediaTek specific code instead of a cleanup.