Replace MediaTek version of Binder with official Binder.
The MediaTek version of Binder provides two extra features over the
official Binder. First, it offers extra debug information through
debugfs files. And second, it modifies the priority of the thread that
replies to a transaction to use the same priority as the thread that it
is replying to.

However, when used on Android 4.4, the code to change the thread
priority causes a kernel panic due to a null pointer dereference (line
2547 of binder.c). As the additional features provided by the MediaTek
version of Binder are not needed it was replaced by its equivalent
official version, that is, the binder.c file found on the official
Android kernel at commit 633c6f1b21 ("Merge commit 'v3.4.5' into
android-3.4"). This Binder version works on Android 4.4 as long as
SELinux support is disabled; further patches, that can be got from the
official Android kernel, would be needed if enabled.

Note that, despite this change, this kernel can still be used on an
official Android 4.2 Fairphone 1 ROM; MediaTek services do not seem to
use the extra debug information of Binder on release builds (as debugfs
filesystem is not mounted), and not inheriting the priority does not
seem to cause any problem either.