Enable ext4 security labels in defconfig for Fairphone 1 recovery.
When the updater script inside an Android OTA package uses the
"set_metadata" or "set_metadata_recursive" commands the kernel of the
recovery used to install the package must support ext4 security labels.
Otherwise setting the metadata would fail which, in turn, would cause
the package installation to fail too.

Note, however, that if the kernel of the installed system does not
support the ext4 security labels those set during the installation would
be simply ignored; if the system works properly or not without that
support depends on other factors, like whether SELinux is enabled or
not (as it would need to access to the security labels).